Ref: auth page refactored #129
