+++ src/app/admin/page.jsx
@@ -17,7 +17,6 @@ import {
 } from '@nextui-org/react';
 import { IoSearch } from 'react-icons/io5';
 
-
 import AdminHeader from '@/components/admin/AdminHeader';
 import UserDetailsModal from '@/components/admin/UserDetailModal';
