--- src/app/auth/layout.js
@@ -1,14 +1,0 @@
-//import { Suspense } from "react";
-
-export const metadata = {
-    title: "Auth",
-    description: "Authentication and authorization management platform",
-};
-
-export default function AuthLayout({ children }) {
-    return (
-        <>
-            {children}
-        </>
-    );
-}
