+++ src/app/auth/signin/[social]/callback/page.jsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { GoogleAuth } from './GoogleAuth';
+import { GoogleAuth } from '@/services/auth/signin/google/GoogleAuth';
 
 export default function Page() {
   return <GoogleAuth />;
