+++ src/app/auth/signin/layout.js
@@ -1,0 +1,14 @@
+//import { Suspense } from "react";
+
+export const metadata = {
+    title: "SignIn",
+    description: "SignIn to your account",
+};
+
+export default function SignInLayout({ children }) {
+    return (
+        <>
+            {children}
+        </>
+    );
+}
