+++ src/app/auth/signup/layout.js
@@ -1,0 +1,14 @@
+//import { Suspense } from "react";
+
+export const metadata = {
+    title: "SignUp",
+    description: "SignUp to your account",
+};
+
+export default function SignUpLayout({ children }) {
+    return (
+        <>
+            {children}
+        </>
+    );
+}
