+++ src/app/auth/signup/page.jsx
@@ -4,9 +4,10 @@ import { useState, useEffect } from "react";
 import { Button } from "@nextui-org/react";
 import { useRouter } from "next/navigation";
 
-import ProfileInfoForm from "./components/ProfileInfoForm";
-import AdditionalInfoForm from "./components/AdditionalInfoForm";
-import usePasswordValidation from "./hooks/usePasswordValidation";
+import ProfileInfoForm from "@/components/auth/signup/ProfileInfoForm";
+import AdditionalInfoForm from "@/components/auth/signup/AdditionalInfoForm";
+
+import usePasswordValidation from "@/hooks/usePasswordValidation";
 import { useAuthenticatedApi } from '@/hooks/useAuthenticatedApi.js';
 
 export default function Signup() {
