+++ src/app/main/layout.js
@@ -1,0 +1,14 @@
+//import { Suspense } from "react";
+
+export const metadata = {
+    title: "Home",
+    description: "Home management and participation platform",
+};
+
+export default function HomeLayout({ children }) {
+    return (
+        <>
+            {children}
+        </>
+    );
+}
