+++ src/components/auth/screen/AuthLogin.jsx
@@ -4,7 +4,7 @@ import { Button, Form, Input, Spinner } from '@nextui-org/react';
 import Image from 'next/image';
 import Link from 'next/link';
 
-import googleIcon from '@public/icons/logo.png';
+import googleIcon from '@public/icons/logo/google.png';
 
 export default function AuthLogin({ router, onSubmit, errors, password, setPassword, setErrors, handleGoogleLogin, handleFindIdClick, handleResetPasswordClick }) {
   return (
