+++ src/components/landing/ActivitiesSection.jsx
