+++ src/components/landing/PartnersSection.jsx
