+++ src/components/landing/StatsSection.jsx
