+++ src/components/landing/StudySection.jsx
